Agent/agent-ai/requirements.txt
a2a==0.44
a2a-sdk==0.3.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==4.0.0
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.1
itemloaders==1.3.2
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith==0.4.8
lxml==6.0.0
marshmallow==3.26.1
mem0ai==0.1.115
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
parsel==1.10.0
portalocker==3.2.0
posthog==6.3.2
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
queuelib==1.8.0
regex==2025.7.31
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tldextract==5.3.0
tqdm==4.67.1
Twisted==25.5.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
w3lib==2.3.1
xxhash==3.5.0
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
langchain_mcp_adapters=0.1.9
